Add note about build's AWS credentials
Ideally these credentials should only be used for uploading the
artifact. Granting more widespread AWS access to the credentials
Travis controls is often not a good idea and attempting to do so
should trigger a conversation. I've added a few comments to try and
make it clear the intent here is just to allow Travis to upload the
build's artifacts.